using endpoints library for generation of openAPI doc with swagger
