+++ build.sbt
@@ -34,12 +34,12 @@ libraryDependencies  ++=  Seq(
   "io.circe"                        %% "circe-parser"              % circeVersion,
   "io.circe"                        %% "circe-generic"             % circeVersion,
   "io.circe"                        %% "circe-generic-extras"      % circeVersion,
-  "de.heikoseeberger" %% "akka-http-circe" % "1.23.0" exclude("com.typesafe.akka", "akka-http"),
-  "org.julienrf" %% "endpoints-algebra" % endpointsVersion,
-  "org.julienrf" %% "endpoints-openapi" % endpointsVersion,
-  "org.julienrf" %% "endpoints-json-schema-generic" % endpointsVersion,
-  "org.julienrf" %% "endpoints-akka-http-server" % endpointsVersion,
-  "com.chuusai" %% "shapeless" % "2.3.3",
+  "de.heikoseeberger"               %% "akka-http-circe"           % "1.23.0" exclude("com.typesafe.akka", "akka-http"),
+  "org.julienrf"                    %% "endpoints-algebra"         % endpointsVersion,
+  "org.julienrf"                    %% "endpoints-openapi"         % endpointsVersion,
+  "org.julienrf"                    %% "endpoints-json-schema-generic" % endpointsVersion,
+  "org.julienrf"                    %% "endpoints-akka-http-server" % endpointsVersion,
+  "com.chuusai"                     %% "shapeless"                 % "2.3.3",
   "de.heikoseeberger"               %% "akka-http-jackson"         % "1.23.0",
   "com.typesafe.slick"              %% "slick"                     % slickVersion exclude("org.reactivestreams", "reactive-streams") exclude("com.typesafe", "config") exclude("org.slf4j", "slf4j-api"),
   "com.typesafe.slick"              %% "slick-hikaricp"            % slickVersion exclude("org.slf4j", "slf4j-api"),
