+++ src/main/scala/tech/cryptonomic/conseil/routes/openapi/OpenApiDoc.scala
@@ -1,22 +1,22 @@
 package tech.cryptonomic.conseil.routes.openapi
 
 import endpoints.algebra.Documentation
-import endpoints.{InvariantFunctor, openapi}
 import endpoints.openapi.model._
+import endpoints.{InvariantFunctor, openapi}
 import io.circe.Json
 import io.circe.syntax._
 
 object OpenApiDoc
-    extends Endpoints
+  extends DataEndpoints
     with openapi.Endpoints
     with openapi.JsonSchemaEntities
     with openapi.BasicAuthentication {
 
-  def openapi: OpenApi = openApi(Info("Conseil API", "0.0.1"))(queryEndpoint, blocksEndpoint, blocksHeadEndpoint, blockByHashEndpoint,
-    accountsEndpoint, accountByIdEndpoint, operationGroupsEndpoint, operationGroupByIdEndpoint, avgFeesEndpoint, operationsEndpoint)
-
   def openapiJson: Json =
     openapi.asJson
+
+  def openapi: OpenApi = openApi(Info("Conseil API", "0.0.1"))(queryEndpoint, blocksEndpoint, blocksHeadEndpoint, blockByHashEndpoint,
+    accountsEndpoint, accountByIdEndpoint, operationGroupsEndpoint, operationGroupByIdEndpoint, avgFeesEndpoint, operationsEndpoint)
 
   def validated[A](response: List[OpenApiDoc.DocumentedResponse], invalidDocs: Documentation): List[OpenApiDoc.DocumentedResponse] =
     response :+ OpenApiDoc.DocumentedResponse(
@@ -27,9 +27,9 @@ object OpenApiDoc
       )
     )
 
-  override implicit def anySchema: DocumentedJsonSchema = DocumentedJsonSchema.Primitive("Any")
+  override implicit def anySchema: DocumentedJsonSchema = DocumentedJsonSchema.Primitive("Any - not yet supported")
 
-  override implicit def queryResponseSchema: DocumentedJsonSchema = DocumentedJsonSchema.Primitive("Any")
+  override implicit def queryResponseSchema: DocumentedJsonSchema = DocumentedJsonSchema.Primitive("Any - not yet supported")
 
   override def qsList[A: QueryStringParam](name: String, docs: Option[String]): DocumentedQueryString = new DocumentedQueryString(
     List(
@@ -41,7 +41,7 @@ object OpenApiDoc
     def xmap[From, To](f: OpenApiDoc.DocumentedQueryString, map: From => To, contramap: To => From): OpenApiDoc.DocumentedQueryString = f
   }
 
-  override implicit def timestampSchema: DocumentedJsonSchema = DocumentedJsonSchema.Primitive("Long")
+  override implicit def timestampSchema: DocumentedJsonSchema = DocumentedJsonSchema.Primitive("integer")
 
-  override implicit def blocksByHashSchema: DocumentedJsonSchema = DocumentedJsonSchema.Primitive("Any")
+  override implicit def blocksByHashSchema: DocumentedJsonSchema = DocumentedJsonSchema.Primitive("Any - not yet supported")
 }
