+++ src/main/scala/tech/cryptonomic/conseil/routes/openapi/Validation.scala
@@ -5,7 +5,5 @@ import endpoints.algebra.Documentation
 import tech.cryptonomic.conseil.generic.chain.DataTypes.QueryValidationError
 
 trait Validation extends algebra.Responses {
-
-    def validated[A](response: Response[A], invalidDocs: Documentation): Response[Either[List[QueryValidationError], A]]
-
+  def validated[A](response: Response[A], invalidDocs: Documentation): Response[Either[List[QueryValidationError], A]]
 }
